rrr: convert rpc/heartbeat to module (rrr.heartbeat)
heartbeat.cpp is the rrr.heartbeat module interface unit;
heartbeat.hpp deleted. heartbeat_time_us() free fn, HeartbeatConfig
POD with 3 presets (aggressive / relaxed / disabled), and
HeartbeatManager (rusty::Cell-backed state + on_timeout callback).

GMF gotcha: needed explicit `#include <cstdlib>` for std::abort that
rusty/function.hpp references internally. Without it, the textual
include path inside the GMF fails with "no member named 'abort' in
namespace 'std'".

Includers (2) updated. 23 modules total.
